refactor(transform): isolate snapshot replay
